--- src/test/java/com/ericgrandt/services/EconomyServiceImplTest.java
@@ -1,4 +1,0 @@
-package com.ericgrandt.services;
-
-public class EconomyServiceImplTest {
-}
